Fixed Blue Peg & Changed Stronghold
I added an RSS link for Blue Peg / Ping Peg and changed the Stronghold
Games banner.  This involved deleting the old banner image and uploading
a new one as well as minor updates in the index.php and styles.css files
to account for the addition of the RSS feed link.